// Sportify/src/models/achievement.ts
export interface UserAchievement {
  title: string;
  description: string;
  isAutoGenerated: boolean;
}

export interface FullAchievement extends UserAchievement {
  achievementId: number;
}

export interface CreateAchievementRequest {
  title: string;
  description: string;
  isAutoGenerated: boolean;
}

export interface AssignAchievementRequest {
  userId: string;
  achievementId: string;
  eventId?: string;
  awardedByAdminId: string;
}